+++ elements/cache/prod/smarty_cache/5626b3e68ade1205c06093906c72edd68af98ca7.index.tpl.php
@@ -1,11 +1,11 @@
 <?php
-/* Smarty version 3.1.31, created on 2017-10-09 02:41:30
+/* Smarty version 3.1.31, created on 2017-10-17 13:55:30
   from "/Applications/MAMP/htdocs/iumio-framework/apps/HelloApp/Front/views/index.tpl" */
 
 /* @var Smarty_Internal_Template $_smarty_tpl */
 if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
   'version' => '3.1.31',
-  'unifunc' => 'content_59dac5bad8c293_81153224',
+  'unifunc' => 'content_59e5efb29bfc22_89764571',
   'has_nocache_code' => false,
   'file_dependency' => 
   array (
@@ -24,7 +24,7 @@ if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
   ),
   'cache_lifetime' => 3600,
 ),true)) {
-function content_59dac5bad8c293_81153224 (Smarty_Internal_Template $_smarty_tpl) {
+function content_59e5efb29bfc22_89764571 (Smarty_Internal_Template $_smarty_tpl) {
 ?>
 <!DOCTYPE HTML>
 <!--