+++ elements/compiled/dev/smarty_compiled/11dc170af87665ac57c82823b3e80d45257bae06_0.file.smartymanager.tpl.php
@@ -1,17 +1,17 @@
 <?php
-/* Smarty version 3.1.31, created on 2017-10-17 14:50:26
-  from "/Applications/MAMP/htdocs/iumio-framework/vendor/iumio_framework/php/BaseApps/ManagerApp/Front/views/compilemanager.tpl" */
+/* Smarty version 3.1.31, created on 2017-10-17 14:50:28
+  from "/Applications/MAMP/htdocs/iumio-framework/vendor/iumio_framework/php/BaseApps/ManagerApp/Front/views/smartymanager.tpl" */
 
 /* @var Smarty_Internal_Template $_smarty_tpl */
 if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
   'version' => '3.1.31',
-  'unifunc' => 'content_59e5fc924e1025_88394655',
+  'unifunc' => 'content_59e5fc94d15880_08228617',
   'has_nocache_code' => false,
   'file_dependency' => 
   array (
-    '2a4ae61a97df933e206f99933f067fa5afe8eef5' => 
+    '11dc170af87665ac57c82823b3e80d45257bae06' => 
     array (
-      0 => '/Applications/MAMP/htdocs/iumio-framework/vendor/iumio_framework/php/BaseApps/ManagerApp/Front/views/compilemanager.tpl',
+      0 => '/Applications/MAMP/htdocs/iumio-framework/vendor/iumio_framework/php/BaseApps/ManagerApp/Front/views/smartymanager.tpl',
       1 => 1507506917,
       2 => 'file',
     ),
@@ -23,22 +23,22 @@ if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
     'file:partials/footer.tpl' => 1,
   ),
 ),false)) {
-function content_59e5fc924e1025_88394655 (Smarty_Internal_Template $_smarty_tpl) {
+function content_59e5fc94d15880_08228617 (Smarty_Internal_Template $_smarty_tpl) {
 $_smarty_tpl->_loadInheritance();
 $_smarty_tpl->inheritance->init($_smarty_tpl, true);
 ?>
 
 <?php 
-$_smarty_tpl->inheritance->instanceBlock($_smarty_tpl, 'Block_82636283359e5fc91762668_30261255', "principal");
+$_smarty_tpl->inheritance->instanceBlock($_smarty_tpl, 'Block_111328772859e5fc94049d54_68876580', "principal");
 $_smarty_tpl->inheritance->endChild($_smarty_tpl, 'template.tpl');
 }
 /* {block "principal"} */
-class Block_82636283359e5fc91762668_30261255 extends Smarty_Internal_Block
+class Block_111328772859e5fc94049d54_68876580 extends Smarty_Internal_Block
 {
 public $subBlocks = array (
   'principal' => 
   array (
-    0 => 'Block_82636283359e5fc91762668_30261255',
+    0 => 'Block_111328772859e5fc94049d54_68876580',
   ),
 );
 public function callBlock(Smarty_Internal_Template $_smarty_tpl) {
@@ -55,7 +55,7 @@ public function callBlock(Smarty_Internal_Template $_smarty_tpl) {
                         <?php $_smarty_tpl->_subTemplateRender('file:partials/toogle.tpl', $_smarty_tpl->cache_id, $_smarty_tpl->compile_id, 0, $_smarty_tpl->cache_lifetime, array(), 0, false);
 ?>
 
-                        <a class="navbar-brand" href="#">Compiled Manager</a>
+                        <a class="navbar-brand" href="#">Smarty Manager</a>
                     </div>
                     <div class="collapse navbar-collapse">
                         <ul class="nav navbar-nav navbar-left">
@@ -70,38 +70,21 @@ public function callBlock(Smarty_Internal_Template $_smarty_tpl) {
                         <div class="col-md-12">
                             <div class="card">
                                 <div class="header">
-                                    <h4 class="title">Options</h4>
-                                    <p class="category">Actions to clear compiled files</p>
-                                </div>
-                                <div class="content">
-                                    <div class="row center-block text-center manager-options">
-                                        <div class="col-md-12">
-                                            <a class="btn-default btn clearcompile"  attr-href="<?php echo iumioFramework\Core\Additionnal\Template\ViewBasePlugin::route(array('name'=>'iumio_manager_compile_manager_remove_all'),$_smarty_tpl);?>
-">Clear all compiled</a>
-                                        </div>
-
-                                    </div>
-                                </div>
-                            </div>
-                        </div>
-                        <div class="col-md-12">
-                            <div class="card">
-                                <div class="header">
-                                    <h4 class="title">Compiled list</h4>
-                                    <p class="category">Referer to compilation directory</p>
+                                    <h4 class="title">List of smarty configurations</h4>
+                                    <p class="category">Referer to smarty.json</p>
                                 </div>
                                 <div class="content table-responsive table-full-width">
                                     <table class="table table-hover table-striped">
                                         <thead>
-                                        <th>Directory name</th>
-                                        <th>Path</th>
-                                        <th>Size</th>
-                                        <th>Permissions</th>
-                                        <th>Status</th>
-                                        <th>Clear</th>
+                                        <th>Config name</th>
+                                        <th>Debug</th>
+                                        <th>Cache</th>
+                                        <th>Console</th>
+                                        <th>Edit</th>
                                         </thead>
-                                        <tbody class="getAllEnvCompile" attr-href="<?php echo iumioFramework\Core\Additionnal\Template\ViewBasePlugin::route(array('name'=>'iumio_manager_compile_manager_get_all'),$_smarty_tpl);?>
+                                        <tbody class="smartyconfigs" attr-href="<?php echo iumioFramework\Core\Additionnal\Template\ViewBasePlugin::route(array('name'=>'iumio_manager_smarty_manager_get_all'),$_smarty_tpl);?>
 ">
+
                                         </tbody>
                                     </table>
 
@@ -109,10 +92,9 @@ public function callBlock(Smarty_Internal_Template $_smarty_tpl) {
                             </div>
                         </div>
                     </div>
-
-
                 </div>
             </div>
+
             <?php $_smarty_tpl->_subTemplateRender('file:partials/footer.tpl', $_smarty_tpl->cache_id, $_smarty_tpl->compile_id, 0, $_smarty_tpl->cache_lifetime, array(), 0, false);
 ?>